add test case for DeleteBookId for non-existent book id, modify Controller to pass test

+++ BookManagerApi/Controllers/BookManagerController.cs
@@ -61,6 +61,9 @@ namespace BookManagerApi.Controllers
         [HttpDelete("{id}")]
         public IActionResult DeleteBookById(long id)
         {
+            if (!_bookManagementService.BookExists(id))
+                return NotFound();
+
             _bookManagementService.Delete(id);
             return NoContent();
         }
